install.py: add --dry-run option

+++ install.py
@@ -2,11 +2,16 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+import argparse
 import datetime
 import os
 
 
 def main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument('-n', '--dry-run', action='store_true')
+    args = parser.parse_args()
+
     timestamp = datetime.datetime.now().strftime('%Y%m%d%H%M%S')
     configdir = os.path.abspath(os.path.dirname(__file__))
     homedir = os.path.expanduser('~')
@@ -20,11 +25,16 @@ def main():
 
         source = os.path.join(homedir, filename)
         target = os.path.join(configdir, filename)
+
+        print('{0} -> {1}'.format(source, target))
+        if args.dry_run:
+            continue
+
         if os.path.islink(source):
             os.unlink(source)
         if os.path.exists(source):
             os.rename(source, '%s.%s' % (source, timestamp))
-        print('{0} -> {1}'.format(source, target))
+
         if 'HARDLINK' in comment:
             os.link(target, source)
         else:
